rename: FilterInfoButton.js to FilterInfoButton.tsx
